--- Misoten8/Assets/ImportAssets/MusicPlayer(Reaktor)/Script/Reaktion/Reaktor.meta
@@ -1,7 +1,0 @@
-fileFormatVersion: 2
-guid: e8d657bae5a434fcb959b69b3b33f3f8
-folderAsset: yes
-DefaultImporter:
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
