--- Misoten8/Assets/ImportAssets/TextFx/Editor/Icons.meta
@@ -1,7 +1,0 @@
-fileFormatVersion: 2
-guid: 0b1f9f84468ff344196894b9cd14656e
-folderAsset: yes
-DefaultImporter:
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
