+++ Misoten8/Assets/Models/Materials.meta
@@ -1,7 +1,7 @@
 fileFormatVersion: 2
-guid: 87150f810657f7849baeded4e826aea0
+guid: ec76e55dd96250c4ab3fcf15b7c3569c
 folderAsset: yes
-timeCreated: 1512957344
+timeCreated: 1512622337
 licenseType: Free
 DefaultImporter:
   userData:
