+++ Misoten8/Assets/Models/Materials/lambert2.mat.meta
@@ -1,0 +1,9 @@
+fileFormatVersion: 2
+guid: f6d856dc7c814b844bbaa049df049757
+timeCreated: 1513068065
+licenseType: Free
+NativeFormatImporter:
+  mainObjectFileID: 2100000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
