--- Misoten8/Assets/Plugins/DanceArea.meta
@@ -1,9 +1,0 @@
-fileFormatVersion: 2
-guid: 8e9742981cfd38c488df9dcf25a471c2
-folderAsset: yes
-timeCreated: 1510462475
-licenseType: Free
-DefaultImporter:
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
