--- Misoten8/Assets/Plugins/MusicPlayer(Reaktor)/Script/Reaktion.meta
@@ -1,9 +1,0 @@
-fileFormatVersion: 2
-guid: 3169e0ddbab4ed1418ee65fb1dc9482c
-folderAsset: yes
-timeCreated: 1510462475
-licenseType: Free
-DefaultImporter:
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
