+++ Misoten8/Assets/Resources/Models/Materials.meta
@@ -1,7 +1,7 @@
 fileFormatVersion: 2
-guid: 0f02f82e2dc2b1b4580360ec64744e16
+guid: b672dc6dd640bb048bb92d71f7c28406
 folderAsset: yes
-timeCreated: 1512957344
+timeCreated: 1513068087
 licenseType: Free
 DefaultImporter:
   userData:
