+++ Misoten8/Assets/Resources/Models/Materials/lambert1.mat.meta
@@ -1,0 +1,9 @@
+fileFormatVersion: 2
+guid: 99e3963b731c3d44c8a2ab6c3c96e14b
+timeCreated: 1513068162
+licenseType: Free
+NativeFormatImporter:
+  mainObjectFileID: 2100000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
