+++ Misoten8/Assets/Resources/Models/Materials/lambert2.mat.meta
@@ -1,0 +1,9 @@
+fileFormatVersion: 2
+guid: c4e55f12b1f23384199bb98340717352
+timeCreated: 1513068162
+licenseType: Free
+NativeFormatImporter:
+  mainObjectFileID: 2100000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
