+++ Misoten8/Assets/Resources/Models/Materials/lambert3.mat.meta
@@ -1,0 +1,9 @@
+fileFormatVersion: 2
+guid: cd52d326fb8d8e3408e9ea39e1b435bd
+timeCreated: 1513068162
+licenseType: Free
+NativeFormatImporter:
+  mainObjectFileID: 2100000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
