+++ Misoten8/Assets/Resources/Models/Materials/lambert4.mat.meta
@@ -1,0 +1,9 @@
+fileFormatVersion: 2
+guid: f06bf0759c107324884b7859e9b47247
+timeCreated: 1513068166
+licenseType: Free
+NativeFormatImporter:
+  mainObjectFileID: 2100000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
