+++ Misoten8/Assets/Resources/Models/Materials/lambert5.mat.meta
@@ -1,0 +1,9 @@
+fileFormatVersion: 2
+guid: 8d8030eac92e3e8419cea738bbde9e64
+timeCreated: 1513068165
+licenseType: Free
+NativeFormatImporter:
+  mainObjectFileID: 2100000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
