+++ Misoten8/Assets/Resources/Models/Materials/lambert6.mat.meta
@@ -1,0 +1,9 @@
+fileFormatVersion: 2
+guid: 087f3aea822375e47adaa415e0a41ce0
+timeCreated: 1513068165
+licenseType: Free
+NativeFormatImporter:
+  mainObjectFileID: 2100000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
