+++ Misoten8/Assets/Resources/Models/Materials/player1_body.mat.meta
@@ -1,0 +1,9 @@
+fileFormatVersion: 2
+guid: a7ac2c5890a131e49bd8180c5e7c5ef8
+timeCreated: 1513068087
+licenseType: Free
+NativeFormatImporter:
+  mainObjectFileID: 2100000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
