+++ Misoten8/Assets/Resources/Models/Materials/player1_head.mat.meta
@@ -1,0 +1,9 @@
+fileFormatVersion: 2
+guid: a05c3002505d40246896a782330933b6
+timeCreated: 1513068087
+licenseType: Free
+NativeFormatImporter:
+  mainObjectFileID: 2100000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
