--- Misoten8/Assets/Scripts/Input/Wiimote/WiiTest.cs.meta
@@ -1,12 +1,0 @@
-fileFormatVersion: 2
-guid: 12f5cb0485f42b24dae4250061339e5b
-timeCreated: 1507175118
-licenseType: Free
-MonoImporter:
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
