+++ Misoten8/Assets/Scripts/Input/Wiimote/Wiimote.cs.meta
@@ -1,6 +1,6 @@
 fileFormatVersion: 2
-guid: 5463cc53d9e56334aabf1ad0fe825524
-timeCreated: 1429997858
+guid: 1c00b625fced1fb49a709081ba068c4f
+timeCreated: 1513069077
 licenseType: Free
 MonoImporter:
   serializedVersion: 2
