+++ Misoten8/Assets/Scripts/Input/Wiimote/WiimoteManager.cs.meta
@@ -1,6 +1,6 @@
 fileFormatVersion: 2
-guid: 52b7084b4645b2248a64b3e8fb3dd28e
-timeCreated: 1428076367
+guid: dbd9cf3367689d140b5bebd3a76a8b6c
+timeCreated: 1513069077
 licenseType: Free
 MonoImporter:
   serializedVersion: 2
