--- Misoten8/Assets/Scripts/Wiimote/WiimoteData.meta
@@ -1,9 +1,0 @@
-fileFormatVersion: 2
-guid: 5985c751389549545bff3a8f3b580b43
-folderAsset: yes
-timeCreated: 1512957344
-licenseType: Free
-DefaultImporter:
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
